+++ packages/sdk/src/plt/index.ts
@@ -11,4 +11,5 @@ export * as TokenAmount from './TokenAmount.js';
 export * as Token from './Token.js';
 export * as Cbor from './Cbor.js';
 export * as CborMemo from './CborMemo.js';
+export * as CborAccountAddress from './CborAccountAddress.js';
 export * as TokenHolder from './TokenHolder.js';